requirements.txt
django>=4.2.0,<=5.0.0
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2-binary
uritemplate
django-rest-swagger
djangorestframework_simplejwt
django-cockroachdb==5.0.*
dj_database_url
drf-yasg